pyproject.toml
[tool.poetry]
name = "flipkart-login-playwright-test"
version = "0.1.0"
description = "this project is demo for playwright python test where we write tests!"
authors = ["kevalpansuriya <kevalpansuriya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest-playwright = "^0.4.4"
pytest-xdist = "^3.5.0"
pytest-html = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
